Adding basic dungeon stub
Implements most of the dungeon messages from the Evernote notebook. I'm
not perfectly how much mediation is supposed to be done here since I saw
more recent drawings where the player is talking directly to a room.

Also, need to sort out with Andrew if we're defining stuff in #records
vs yaml configs