Participle and verb master list eraticated!
the verb type is now a synonym for an atom, which is good for readability
participle no longer exists. While this may be awkward to refer to past events in present tense, it avoids the need for conversion and the complexities of the English language. Conversion would require some master conversion function, which would require a master list, which in poor design and was removed.