Made player and ai "inherit" a thing module.
This means that receiveEventNotification can be called on the module 'thing,' without knowing what kind of thing it is. the function was removed from player, and inherited.
room was updated accordingly.
Additionally, there was a conflict with the thing() type in room, which was renamed thing_proc since it is it a union of procs.